IRC10.1/Parcial 2/Producto/requirements.txt
flask
flask-wtf
paramiko
ansible
pyyaml